add: отправка запроса на добавление нового пользователя
